split mirror.py
